python/python_django/app_blog2/requirements.txt
asgiref==3.3.1
astroid==2.4.2
backcall==0.2.0
boto3==1.17.30
botocore==1.20.30
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
dj-static==0.0.6
Django==3.1.4
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-storages==1.11.1
django-js-asset==1.2.2
gunicorn==20.0.4
idna==2.10
ipykernel==5.3.4
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
mysql-connector-python==8.0.22
mysqlclient==2.0.3
numpy==1.19.4
openpyxl==3.0.6
pandas==1.1.5
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
protobuf==3.14.0
psycopg2==2.8.6
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-xlsx==0.6.0
Pygments==2.7.3
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.4
pyzmq==20.0.0
requests==2.25.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
urllib3==1.26.2
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
XlsxWriter==1.3.7
